rps/requirements.txt
e3db
python-dotenv
dateutil
parser